+++ objects/pursue503/chapter5/src/DiscountCondition.java
@@ -1,27 +1,3 @@
-import java.time.DayOfWeek;
-import java.time.LocalTime;
-
-public class DiscountCondition {
-    private DisCountConditionType type;
-    private int sequence;
-    private DayOfWeek dayOfWeek;
-    private LocalTime startTime;
-    private LocalTime endTime;
-
-    public boolean isSatisfiedBy(Screening screening) {
-        if (type == DisCountConditionType.PERIOD) {
-            return isSatisfiedByPeriod(screening);
-        }
-        return isSatisfiedBySequence(screening);
-    }
-
-    private boolean isSatisfiedByPeriod(Screening screening) {
-        return dayOfWeek.equals(screening.getWhenScreened().getDayOfWeek()) &&
-                startTime.compareTo(screening.getWhenScreened().toLocalTime()) <= 0 &&
-                endTime.isAfter(screening.getWhenScreened().toLocalTime());
-    }
-
-    private boolean isSatisfiedBySequence(Screening screening) {
-        return sequence == screening.getSequence();
-    }
+public interface DiscountCondition {
+    boolean isSatisfiedBy(Screening screening);
 }
