+++ objects/pursue503/chapter5/src/Movie.java
@@ -8,8 +8,6 @@ public class Movie {
     private List<DiscountCondition> discountConditions;
 
     private MovieType movieType;
-    private Money discountAmount;
-    private double discountPercent;
 
     public Money calculateMovieFee(Screening screening) {
         if (isDiscountable(screening)) {
