+++ objects/pursue503/chapter5/src/SequenceCondition.java
@@ -1,0 +1,12 @@
+public class SequenceCondition implements DiscountCondition {
+    private int sequence;
+
+    public SequenceCondition(int sequence) {
+        this.sequence = sequence;
+    }
+
+    @Override
+    public boolean isSatisfiedBy(Screening screening) {
+        return screening.isSequence(sequence);
+    }
+}
